requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-allauth==0.57.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
psycopg2-binary==2.9.10
Pillow==12.0.0
